refactor(wave_compute): migrate to platform adapter fetchIssue + promote to lib/
Closes #302
